move binding files to their respective directories
