+++ rust/examples/ruapu.rs
