+++ rust/src/lib.rs
