+++ rust/src/ruapu-binding.c
@@ -1,0 +1,2 @@
+#define RUAPU_IMPLEMENTATION
+#include "../../ruapu.h"
